requirements.txt
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-cors-headers==3.0.2
django-filter==2.2.0
django-rest-auth==0.9.5
django-suit==0.2.26
django-templated-mail==1.1.1
djangorestframework==3.10.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djoser==2.0.1
httpie==1.0.2
idna==2.8
oauthlib==3.1.0
Pygments==2.4.2
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
